refactor: 조건부 렌더링 추상화 및 파일 분리
