+++ frontend/src/hooks/useModal.ts
@@ -1,0 +1,11 @@
+import { useState } from 'react';
+
+export const useModal = () => {
+  const [isOpen, setIsOpen] = useState(false);
+
+  const openModal = () => setIsOpen(true);
+  const closeModal = () => setIsOpen(false);
+  const toggleModal = () => setIsOpen((prev) => !prev);
+
+  return { isOpen, openModal, closeModal, toggleModal };
+};
